+++ test/functional/database-schema/column-types/sqlite/column-types-sqlite.ts
@@ -1,5 +1,6 @@
 import "reflect-metadata";
 import {Post} from "./entity/Post";
+import {getManager} from "../../../../../src";
 import {Connection} from "../../../../../src/connection/Connection";
 import {closeTestingConnections, createTestingConnections, reloadTestingDatabases} from "../../../../utils/test-utils";
 import {PostWithoutTypes} from "./entity/PostWithoutTypes";
@@ -170,4 +171,20 @@ describe("database schema > column types > sqlite", () => {
 
     })));
 
+    it("unconventional datetime formats should be hydrated correctly", () => Promise.all(connections.map(async connection => {
+        const manager = getManager();
+        const postRepository = connection.getRepository(Post);
+        const expectedDatetime = new Date("2001-02-03T04:45:56Z");
+
+        await manager.query(
+            "INSERT INTO post(id, datetime) VALUES(?, ?), (?, ?)",
+            [1, "2001-02-03 04:45:56", 2, "2001-02-03T05:45:56 +01:00"]
+        );
+
+        const loadedPost1 = (await postRepository.findOne(1))!;
+        const loadedPost2 = (await postRepository.findOne(2))!;
+
+        loadedPost1.datetime.should.be.equal(expectedDatetime);
+        loadedPost2.datetime.should.be.equal(expectedDatetime);
+    });
 });